+++ src/app/(withHeader)/layout.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 
-import Header from '@/components/common/Header';
-import Sidebar from '@/components/common/Sidebar';
+import { Header } from '@/components/organisms/Header';
+import { Sidebar } from '@/components/organisms/Sidebar';
 
 export default function WithHeaderLayout({
   children,
