+++ src/app/(withHeader)/notices/[id]/page.tsx
@@ -1,10 +1,10 @@
 import { cookies } from 'next/headers';
 
 import { fetcher } from '@/api/fetcher';
-import ImageCarousel from '@/components/common/ImageCarousel';
-import MarkdownViewer from '@/components/common/MarkdownViewer';
 import NotFound from '@/components/common/NotFound';
 import RedirectWithModal from '@/components/common/RedirectWithModal';
+import { ImageCarousel } from '@/components/molecules/ImageCarousel';
+import { MarkdownViewer } from '@/components/molecules/MarkdownViewer';
 import PostCommentForm from '@/components/post/PostCommentForm';
 import PostCommentList from '@/components/post/PostCommentList';
 import PostFileItem from '@/components/post/PostFileItem';
