+++ src/components/atoms/SkeletonBox/index.ts
@@ -1,0 +1,1 @@
+export * from './SkeletonBox';
