--- src/components/common/SkeletonBox.tsx
@@ -1,14 +1,0 @@
-import React, { HTMLAttributes } from 'react';
-
-interface SkeletonBoxProps extends HTMLAttributes<HTMLDivElement> {
-  className?: string;
-}
-
-export default function SkeletonBox({ className = '', ...props }: SkeletonBoxProps) {
-  return (
-    <div
-      className={`bg-gray-200 rounded animate-pulse ${className}`}
-      {...props}
-    />
-  );
-}
