+++ src/components/molecules/ImageCarousel/index.ts
@@ -1,0 +1,1 @@
+export * from './ImageCarousel';
