+++ src/components/molecules/ImageViewer/ImageViewer.tsx
@@ -4,20 +4,22 @@ import { useState } from 'react';
 
 import { SafeImage } from '@/components/atoms/SafeImage';
 
-interface SingleImageProps {
+interface ImageViewerProps {
   imageUrl: string;
+  imageName?: string;
+  className?: string;
 }
 
-export default function SingleImage({ imageUrl }: SingleImageProps) {
+export function ImageViewer({ imageUrl, imageName, className }: ImageViewerProps) {
   const [modalImage, setModalImage] = useState<string | null>(null);
 
   return (
-    <div>
+    <div className={className}>
       {/* 이미지 */}
       <div className="rounded-xl overflow-hidden mb-3 shadow-sm relative w-full h-48">
         <SafeImage
           src={imageUrl}
-          alt={imageUrl}
+          alt={imageName ?? imageUrl}
           className="w-full h-full object-cover object-center cursor-pointer"
           width={400}
           height={200}
@@ -34,7 +36,7 @@ export default function SingleImage({ imageUrl }: SingleImageProps) {
           <div className="relative w-full max-w-3xl p-4">
             <SafeImage
               src={modalImage}
-              alt={modalImage}
+              alt={imageName ?? modalImage}
               width={1200}
               height={800}
               className="w-full h-auto object-contain"
