+++ src/stores/modalStore.ts
@@ -2,7 +2,7 @@ import { create } from 'zustand';
 
 import { ReactNode } from 'react';
 
-import { ModalAction } from '@/components/common/Modal';
+import { ModalAction } from '@/components/organisms/Modal';
 
 interface OpenModalParams {
   title: string;
